+++ changelog.md
@@ -53,6 +53,8 @@
   Eg: `echo ?.n.typ.kind`
 - Added `minIndex` and `maxIndex` to the `sequtils` module
 
+- Added `os.isRelativeTo` to tell whether a path is relative to another
+
 ## Library changes
 
 - `asyncdispatch.drain` now properly takes into account `selector.hasPendingOperations`